+++ game_engine/physics_engine/colision_object/colision_box.h
@@ -10,25 +10,25 @@
 
 class ColisionBox {
 private:
-    float width;
-    float height;
+    int width;
+    int height;
     Vector2D tl_corner;
     Vector2D tr_corner;
     Vector2D bl_corner;
     Vector2D br_corner;
 
 public:
-    ColisionBox(float width, float height);
-    ColisionBox(float x, float y, float width, float height);
+    ColisionBox(int width, int height);
+    ColisionBox(int x, int y, int width, int height);
 
     Vector2D get_tl_corner() const;
     Vector2D get_tr_corner() const;
     Vector2D get_bl_corner() const;
     Vector2D get_br_corner() const;
-    float get_width() const;
-    float get_height() const;
-    void set_width(float new_width);
-    void set_height(float new_height);
+    int get_width() const;
+    int get_height() const;
+    void set_width(int new_width);
+    void set_height(int new_height);
 
     void update_position(Vector2D new_position);
     bool is_touching(const ColisionBox& other) const;
